# Conflicts:
#	src/SolutionInfo.cs
#	src/Umbraco.Web.UI.Client/package.json
#	src/Umbraco.Web.UI.Client/src/common/services/formhelper.service.js
#	src/Umbraco.Web.UI.Client/src/views/common/overlays/itempicker/itempicker.controller.js
#	src/Umbraco.Web.UI.Client/src/views/components/users/change-password.html
#	src/Umbraco.Web.UI.Client/src/views/content/overlays/publish.controller.js
#	src/Umbraco.Web.UI.Client/src/views/content/overlays/publish.html
#	src/Umbraco.Web.UI.Client/src/views/content/overlays/publishdescendants.controller.js
#	src/Umbraco.Web.UI.Client/src/views/content/overlays/publishdescendants.html
#	src/Umbraco.Web.UI.Client/src/views/datatypes/delete.html
#	src/Umbraco.Web.UI.Client/src/views/documenttypes/create.html
#	src/Umbraco.Web.UI.Client/src/views/logviewer/search.html
#	src/Umbraco.Web.UI.Client/src/views/member/member.edit.controller.js
#	src/Umbraco.Web.UI.Client/src/views/packages/edit.html
#	src/Umbraco.Web.UI.Client/src/views/propertyeditors/dropdownFlexible/dropdownFlexible.controller.js
#	src/Umbraco.Web.UI.Client/src/views/propertyeditors/listview/layouts.prevalues.html
#	src/Umbraco.Web.UI.Client/src/views/propertyeditors/nestedcontent/nestedcontent.controller.js
#	src/Umbraco.Web.UI.Client/src/views/propertyeditors/rte/rte.prevalues.html

+++ src/Umbraco.Web.UI.Client/src/views/propertyeditors/blocklist/umbBlockListPropertyEditor.component.js
@@ -29,13 +29,13 @@
         });
 
     function BlockListController($scope, editorService, clipboardService, localizationService, overlayService, blockEditorService, udiService, serverValidationManager, angularHelper) {
-        
+
         var unsubscribe = [];
         var modelObject;
 
         // Property actions:
-        var copyAllBlocksAction;
-        var deleteAllBlocksAction;
+        var copyAllBlocksAction = null;
+        var deleteAllBlocksAction = null;
 
         var inlineEditing = false;
         var liveEditing = true;
@@ -44,27 +44,27 @@
 
         vm.loading = true;
         vm.currentBlockInFocus = null;
-        vm.setBlockFocus = function(block) {
-            if(vm.currentBlockInFocus !== null) {
+        vm.setBlockFocus = function (block) {
+            if (vm.currentBlockInFocus !== null) {
                 vm.currentBlockInFocus.focus = false;
             }
             vm.currentBlockInFocus = block;
             block.focus = true;
-        }
+        };
+
         vm.supportCopy = clipboardService.isSupported();
 
         vm.layout = []; // The layout object specific to this Block Editor, will be a direct reference from Property Model.
         vm.availableBlockTypes = []; // Available block entries of this property editor.
+        vm.labels = {};
 
-        var labels = {};
-        vm.labels = labels;
         localizationService.localizeMany(["grid_addElement", "content_createEmpty"]).then(function (data) {
-            labels.grid_addElement = data[0];
-            labels.content_createEmpty = data[1];
+            vm.labels.grid_addElement = data[0];
+            vm.labels.content_createEmpty = data[1];
         });
 
         vm.$onInit = function() {
-            if (!vm.umbVariantContent) {
+            if (vm.umbProperty && !vm.umbVariantContent) {// if we dont have vm.umbProperty, it means we are in the DocumentTypeEditor.
                 // not found, then fallback to searching the scope chain, this may be needed when DOM inheritance isn't maintained but scope
                 // inheritance is (i.e.infinite editing)
                 var found = angularHelper.traverseScopeChain($scope, s => s && s.vm && s.vm.constructor.name === "umbVariantContentController");
@@ -84,7 +84,7 @@
             vm.validationLimit = vm.model.config.validationLimit;
 
             vm.listWrapperStyles = {};
-            
+
             if (vm.model.config.maxPropertyWidth) {
                 vm.listWrapperStyles['max-width'] = vm.model.config.maxPropertyWidth;
             }
@@ -100,10 +100,6 @@
             } else if(vm.umbElementEditorContent && vm.umbElementEditorContent.getScope) {
                 scopeOfExistence = vm.umbElementEditorContent.getScope();
             }
-            
-            // Create Model Object, to manage our data for this Block Editor.
-            modelObject = blockEditorService.createModelObject(vm.model.value, vm.model.editor, vm.model.config.blocks, scopeOfExistence, $scope);
-            modelObject.load().then(onLoaded);
 
             copyAllBlocksAction = {
                 labelKey: "clipboard_labelForCopyAllEntries",
@@ -111,28 +107,34 @@
                 icon: "documents",
                 method: requestCopyAllBlocks,
                 isDisabled: true
-            }
+            };
+
             deleteAllBlocksAction = {
                 labelKey: 'clipboard_labelForRemoveAllEntries',
                 labelTokens: [],
                 icon: 'trash',
                 method: requestDeleteAllBlocks,
                 isDisabled: true
-            }
+            };
 
             var propertyActions = [
                 copyAllBlocksAction,
                 deleteAllBlocksAction
             ];
-            
+
             if (vm.umbProperty) {
                 vm.umbProperty.setPropertyActions(propertyActions);
             }
+
+            // Create Model Object, to manage our data for this Block Editor.
+            modelObject = blockEditorService.createModelObject(vm.model.value, vm.model.editor, vm.model.config.blocks, scopeOfExistence, $scope);
+            modelObject.load().then(onLoaded);
+
         };
 
         // Called when we save the value, the server may return an updated data and our value is re-synced
         // we need to deal with that here so that our model values are all in sync so we basically re-initialize.
-        function onServerValueChanged(newVal, oldVal) {   
+        function onServerValueChanged(newVal, oldVal) {
 
             // We need to ensure that the property model value is an object, this is needed for modelObject to recive a reference and keep that updated.
             if (typeof newVal !== 'object' || newVal === null) {// testing if we have null or undefined value or if the value is set to another type than Object.
@@ -142,13 +144,13 @@
             modelObject.update(newVal, $scope);
             onLoaded();
         }
-        
+
         function setDirty() {
             if (vm.propertyForm) {
                 vm.propertyForm.$setDirty();
             }
         }
-        
+
         function onLoaded() {
 
             // Store a reference to the layout model, because we need to maintain this model.
@@ -161,7 +163,7 @@
                 // $block must have the data property to be a valid BlockObject, if not its considered as a destroyed blockObject.
                 if (entry.$block === undefined || entry.$block === null || entry.$block.data === undefined) {
                     var block = getBlockObject(entry);
-    
+
                     // If this entry was not supported by our property-editor it would return 'null'.
                     if (block !== null) {
                         entry.$block = block;
@@ -192,13 +194,38 @@
         }
 
         function getDefaultViewForBlock(block) {
-            
+
+            var defaultViewFolderPath = "views/propertyeditors/blocklist/blocklistentryeditors/";
+
             if (block.config.unsupported === true)
-                return "views/propertyeditors/blocklist/blocklistentryeditors/unsupportedblock/unsupportedblock.editor.html";
+                return defaultViewFolderPath + "unsupportedblock/unsupportedblock.editor.html";
 
             if (inlineEditing === true)
-                return "views/propertyeditors/blocklist/blocklistentryeditors/inlineblock/inlineblock.editor.html";
-            return "views/propertyeditors/blocklist/blocklistentryeditors/labelblock/labelblock.editor.html";
+                return defaultViewFolderPath + "inlineblock/inlineblock.editor.html";
+            return defaultViewFolderPath + "labelblock/labelblock.editor.html";
+        }
+
+        /**
+         * Ensure that the containing content variant languag and current property culture is transfered along
+         * to the scaffolded content object representing this block.
+         * This is required for validation along with ensuring that the umb-property inheritance is constently maintained.
+         * @param {any} content
+         */
+        function ensureCultureData(content) {
+
+            if (!content) return;
+
+            if (vm.umbVariantContent.editor.content.language) {
+                // set the scaffolded content's language to the language of the current editor
+                content.language = vm.umbVariantContent.editor.content.language;
+            }
+            // currently we only ever deal with invariant content for blocks so there's only one
+            content.variants[0].tabs.forEach(tab => {
+                tab.properties.forEach(prop => {
+                    // set the scaffolded property to the culture of the containing property
+                    prop.culture = vm.umbProperty.property.culture;
+                });
+            });
         }
 
         function getBlockObject(entry) {
@@ -206,34 +233,24 @@
 
             if (block === null) return null;
 
-            // ensure that the containing content variant language/culture is transfered along
-            // to the scaffolded content object representing this block. This is required for validation 
-            // along with ensuring that the umb-property inheritance is constently maintained.            
-            if (vm.umbVariantContent.editor.content.language) {
-                block.content.language = vm.umbVariantContent.editor.content.language;
-                // currently we only ever deal with invariant content for blocks so there's only one
-                block.content.variants[0].tabs.forEach(tab => {
-                    tab.properties.forEach(prop => {
-                        prop.culture = vm.umbVariantContent.editor.content.language.culture;
-                    });
-                });
-            }
+            ensureCultureData(block.content);
+            ensureCultureData(block.settings);
 
-            // TODO: Why is there a '/' prefixed? that means this will never work with virtual directories
-            block.view = (block.config.view ? "/" + block.config.view : getDefaultViewForBlock(block));
+            block.view = (block.config.view ? block.config.view : getDefaultViewForBlock(block));
+            block.showValidation = block.config.view ? true : false;
 
             block.hideContentInOverlay = block.config.forceHideContentEditorInOverlay === true || inlineEditing === true;
             block.showSettings = block.config.settingsElementTypeKey != null;
-            block.showCopy = vm.supportCopy && block.config.contentTypeKey != null;// if we have content, otherwise it doesn't make sense to copy.
+            block.showCopy = vm.supportCopy && block.config.contentElementTypeKey != null;// if we have content, otherwise it doesn't make sense to copy.
 
             return block;
         }
 
 
-        function addNewBlock(index, contentTypeKey) {
+        function addNewBlock(index, contentElementTypeKey) {
 
             // Create layout entry. (not added to property model jet.)
-            var layoutEntry = modelObject.create(contentTypeKey);
+            var layoutEntry = modelObject.create(contentElementTypeKey);
             if (layoutEntry === null) {
                 return false;
             }
@@ -243,7 +260,7 @@
             if (blockObject === null) {
                 return false;
             }
-            
+
             // If we reach this line, we are good to add the layoutEntry and blockObject to our models.
 
             // Add the Block Object to our layout entry.
@@ -251,7 +268,7 @@
 
             // add layout entry at the decired location in layout.
             vm.layout.splice(index, 0, layoutEntry);
-            
+
             // lets move focus to this new block.
             vm.setBlockFocus(blockObject);
 
@@ -271,8 +288,12 @@
             var removed = vm.layout.splice(layoutIndex, 1);
             removed.forEach(x => {
                 // remove any server validation errors associated
-                var guid = udiService.getKey(x.contentUdi);                
-                serverValidationManager.removePropertyError(guid, vm.umbProperty.property.culture, vm.umbProperty.property.segment, "", { matchType: "contains" });
+                var guids = [udiService.getKey(x.contentUdi), (x.settingsUdi ? udiService.getKey(x.settingsUdi) : null)];
+                guids.forEach(guid => {
+                    if (guid) {
+                        serverValidationManager.removePropertyError(guid, vm.umbProperty.property.culture, vm.umbProperty.property.segment, "", { matchType: "contains" });
+                    }
+                })
             });
 
             modelObject.removeDataAndDestroyModel(block);
@@ -280,11 +301,11 @@
         }
 
         function deleteAllBlocks() {
-            vm.layout.forEach(entry => {
-                deleteBlock(entry.$block);
-            });
+            while(vm.layout.length) {
+                deleteBlock(vm.layout[0].$block);
+            };
         }
-        
+
         function activateBlock(blockObject) {
             blockObject.active = true;
         }
@@ -317,7 +338,7 @@
             if (blockObject.config.settingsElementTypeKey) {
                 blockSettingsClone = Utilities.copy(blockObject.settings);
             }
-            
+
             var blockEditorModel = {
                 $parentScope: $scope, // pass in a $parentScope, this maintains the scope inheritance in infinite editing
                 $parentForm: parentForm || vm.propertyForm, // pass in a $parentForm, this maintains the FormController hierarchy with the infinite editing view (if it contains a form)
@@ -325,7 +346,6 @@
                 openSettings: openSettings === true,
                 liveEditing: liveEditing,
                 title: blockObject.label,
-                index: blockIndex,
                 view: "views/common/infiniteeditors/blockeditor/blockeditor.html",
                 size: blockObject.config.editorSize || "medium",
                 submit: function(blockEditorModel) {
@@ -366,11 +386,11 @@
 
         vm.showCreateDialog = showCreateDialog;
         function showCreateDialog(createIndex, $event) {
-            
+
             if (vm.blockTypePicker) {
                 return;
             }
-            
+
             if (vm.availableBlockTypes.length === 0) {
                 return;
             }
@@ -403,9 +423,9 @@
                 submit: function(blockPickerModel, mouseEvent) {
                     var added = false;
                     if (blockPickerModel && blockPickerModel.selectedItem) {
-                        added = addNewBlock(createIndex, blockPickerModel.selectedItem.blockConfigModel.contentTypeKey);
+                        added = addNewBlock(createIndex, blockPickerModel.selectedItem.blockConfigModel.contentElementTypeKey);
                     }
-                    
+
                     if(!(mouseEvent.ctrlKey || mouseEvent.metaKey)) {
                         editorService.close();
                         if (added && vm.layout.length > createIndex) {
@@ -439,6 +459,7 @@
                 blockPickerModel.clipboardItems.push(
                     {
                         type: "elementType",
+                        date: entry.date,
                         pasteData: entry.data,
                         blockConfigModel: modelObject.getScaffoldFromAlias(entry.alias),
                         elementTypeModel: {
@@ -448,12 +469,13 @@
                     }
                 );
             });
-            
+
             var arrayEntriesForPaste = clipboardService.retriveEntriesOfType("elementTypeArray", vm.availableContentTypesAliases);
             arrayEntriesForPaste.forEach(function (entry) {
                 blockPickerModel.clipboardItems.push(
                     {
                         type: "elementTypeArray",
+                        date: entry.date,
                         pasteData: entry.data,
                         blockConfigModel: {}, // no block configuration for paste items of elementTypeArray.
                         elementTypeModel: {
@@ -464,6 +486,10 @@
                 );
             });
 
+            blockPickerModel.clipboardItems.sort( (a, b) => {
+                return b.date - a.date
+            });
+
             // open block picker overlay
             editorService.open(blockPickerModel);
 
@@ -472,13 +498,13 @@
         var requestCopyAllBlocks = function() {
 
             var elementTypesToCopy = vm.layout.filter(entry => entry.$block.config.unsupported !== true).map(entry => entry.$block.content);
-            
+
             // list aliases
             var aliases = elementTypesToCopy.map(content => content.contentTypeAlias);
 
             // remove dublicates
             aliases = aliases.filter((item, index) => aliases.indexOf(item) === index);
-            
+
             var contentNodeName = "";
             if(vm.umbVariantContent) {
                 contentNodeName = vm.umbVariantContent.editor.content.name;
@@ -494,7 +520,7 @@
             clipboardService.copy("elementType", block.content.contentTypeAlias, block.content, block.label);
         }
         function requestPasteFromClipboard(index, pasteEntry) {
-            
+
             if (pasteEntry === undefined) {
                 return false;
             }
@@ -512,7 +538,7 @@
 
             // set the BlockObject on our layout entry.
             layoutEntry.$block = blockObject;
-            
+
             // insert layout entry at the decired location in layout.
             vm.layout.splice(index, 0, layoutEntry);
 
@@ -565,11 +591,12 @@
             copyBlock: copyBlock,
             requestDeleteBlock: requestDeleteBlock,
             deleteBlock: deleteBlock,
-            openSettingsForBlock: openSettingsForBlock 
+            openSettingsForBlock: openSettingsForBlock
         }
 
         vm.sortableOptions = {
             axis: "y",
+            containment: "parent",
             cursor: "grabbing",
             handle: ".blockelement__draggable-element",
             cancel: "input,textarea,select,option",
@@ -586,30 +613,31 @@
         function onAmountOfBlocksChanged() {
 
             // enable/disable property actions
-            copyAllBlocksAction.isDisabled = vm.layout.length === 0;
-            deleteAllBlocksAction.isDisabled = vm.layout.length === 0;
+            if (copyAllBlocksAction) {
+                copyAllBlocksAction.isDisabled = vm.layout.length === 0;
+            }
+            if (deleteAllBlocksAction) {
+                deleteAllBlocksAction.isDisabled = vm.layout.length === 0;
+            }
 
             // validate limits:
-            if (vm.propertyForm) {
+            if (vm.propertyForm && vm.validationLimit) {
 
                 var isMinRequirementGood = vm.validationLimit.min === null || vm.layout.length >= vm.validationLimit.min;
                 vm.propertyForm.minCount.$setValidity("minCount", isMinRequirementGood);
-                
+
                 var isMaxRequirementGood = vm.validationLimit.max === null || vm.layout.length <= vm.validationLimit.max;
                 vm.propertyForm.maxCount.$setValidity("maxCount", isMaxRequirementGood);
-                
             }
         }
+
         unsubscribe.push($scope.$watch(() => vm.layout.length, onAmountOfBlocksChanged));
 
-        
         $scope.$on("$destroy", function () {
             for (const subscription of unsubscribe) {
                 subscription();
             }
         });
-
-
     }
 
 })();
